Refactor the node angle calculator stuff
